LPD-59804 Remove individual FDSView components and consolidate into shared ContentAndFile FDSViews to simplify implementation
