--- modules/apps/site/site-cms-site-initializer/src/main/java/com/liferay/site/cms/site/initializer/internal/frontend/data/set/view/cards/SpaceContentsSectionCardsFDSView.java
@@ -1,28 +1,0 @@
-/**
- * SPDX-FileCopyrightText: (c) 2025 Liferay, Inc. https://liferay.com
- * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
- */
-
-package com.liferay.site.cms.site.initializer.internal.frontend.data.set.view.cards;
-
-import com.liferay.frontend.data.set.view.FDSView;
-import com.liferay.site.cms.site.initializer.internal.constants.CMSSiteInitializerFDSNames;
-
-import org.osgi.service.component.annotations.Component;
-
-/**
- * @author Roberto Díaz
- */
-@Component(
-	property = "frontend.data.set.name=" + CMSSiteInitializerFDSNames.SPACE_CONTENTS_SECTION,
-	service = FDSView.class
-)
-public class SpaceContentsSectionCardsFDSView
-	extends BaseContentsSectionCardsFDSView {
-
-	@Override
-	public boolean isDefault() {
-		return false;
-	}
-
-}
